Minor refactoring to ensure unit test names are consistent and filescoped namespaces are used
