+++ src/Domain/Repositories/IEarningsQueryRepository.cs
@@ -1,13 +1,11 @@
 ﻿using SFA.DAS.Funding.ApprenticeshipEarnings.DataTransferObjects;
-using SFA.DAS.Funding.ApprenticeshipEarnings.Domain.ApprenticeshipFunding;
 
-namespace SFA.DAS.Funding.ApprenticeshipEarnings.Domain.Repositories
+namespace SFA.DAS.Funding.ApprenticeshipEarnings.Domain.Repositories;
+
+public interface IEarningsQueryRepository
 {
-    public interface IEarningsQueryRepository
-    {
-        Task Add(Apprenticeship.Apprenticeship apprenticeship);
-        Task Replace(Apprenticeship.Apprenticeship apprenticeship);
-        Task<ProviderEarningsSummary> GetProviderSummary(long ukprn, short academicYear);
-        Task<AcademicYearEarnings> GetAcademicYearEarnings(long ukprn, short academicYear);
-    }
+    Task Add(Apprenticeship.Apprenticeship apprenticeship);
+    Task Replace(Apprenticeship.Apprenticeship apprenticeship);
+    Task<ProviderEarningsSummary> GetProviderSummary(long ukprn, short academicYear);
+    Task<AcademicYearEarnings> GetAcademicYearEarnings(long ukprn, short academicYear);
 }
